add loglik model select
