+++ src/hatchet/hatchet.ini
@@ -154,7 +154,6 @@ diploid = False
 tetraploid = False
 verbosity = 2
 binwise = False
-model_select = "likelihood"
 
 [plot_cn]
 patientnames =
